requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
pydantic==2.11.7
pydantic-settings==2.10.1
jinja2==3.1.6
python-multipart==0.0.20
httpx==0.28.1
PyYAML==6.0.2
Pillow==11.3.0
pytest==8.4.1
